requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
aiohappyeyeballs==2.4.6
    # via aiohttp
aiohttp==3.11.13
    # via geoip2
aiosignal==1.3.2
    # via aiohttp
amqp==5.2.0
    # via kombu
asgiref==3.8.1
    # via
    #   django
    #   django-allauth
    #   django-cors-headers
attrs==25.1.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
billiard==4.2.0
    # via celery
cachetools==5.5.2
    # via google-auth
celery==5.4.0
    # via -r requirements.in
certifi==2023.7.22
    # via
    #   requests
    #   sentry-sdk
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
django==5.0.7
    # via
    #   -r requirements.in
    #   django-allauth
    #   django-cors-headers
    #   django-debug-toolbar
    #   django-filter
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
django-allauth==65.6.0
    # via -r requirements.in
django-cors-headers==4.6.0
    # via -r requirements.in
django-debug-toolbar==4.4.6
    # via -r requirements.in
django-filter==24.2
    # via -r requirements.in
django-ipware==4.0.2
    # via -r requirements.in
djangorestframework==3.15.2
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-spectacular
djangorestframework-simplejwt==5.5.0
    # via -r requirements.in
drf-spectacular==0.28.0
    # via -r requirements.in
factory-boy==3.3.0
    # via -r requirements.in
faker==26.0.0
    # via factory-boy
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
geoip2==5.0.1
    # via -r requirements.in
google-auth==2.38.0
    # via
    #   -r requirements.in
    #   google-auth-oauthlib
google-auth-oauthlib==1.2.1
    # via -r requirements.in
gunicorn==22.0.0
    # via -r requirements.in
idna==3.4
    # via
    #   requests
    #   yarl
inflection==0.5.1
    # via drf-spectacular
iniconfig==2.0.0
    # via pytest
jsonschema==4.23.0
    # via drf-spectacular
jsonschema-specifications==2024.10.1
    # via jsonschema
kombu==5.3.7
    # via celery
maxminddb==2.6.3
    # via geoip2
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
oauthlib==3.2.2
    # via requests-oauthlib
packaging==23.1
    # via
    #   gunicorn
    #   pytest
pluggy==1.5.0
    # via pytest
prompt-toolkit==3.0.47
    # via click-repl
propcache==0.3.0
    # via
    #   aiohttp
    #   yarl
psycopg2-binary==2.9.10
    # via -r requirements.in
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pyjwt==2.8.0
    # via djangorestframework-simplejwt
pytest==8.3.1
    # via
    #   -r requirements.in
    #   pytest-django
pytest-django==4.8.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   celery
    #   faker
pyyaml==6.0.2
    # via drf-spectacular
redis==5.0.7
    # via -r requirements.in
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   -r requirements.in
    #   geoip2
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rpds-py==0.23.1
    # via
    #   jsonschema
    #   referencing
rsa==4.9
    # via google-auth
sentry-sdk==2.10.0
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
sqlparse==0.4.4
    # via
    #   django
    #   django-debug-toolbar
typing-extensions==4.12.2
    # via referencing
tzdata==2024.1
    # via celery
uritemplate==4.1.1
    # via drf-spectacular
urllib3==2.0.4
    # via
    #   requests
    #   sentry-sdk
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
yarl==1.18.3
    # via aiohttp
